+++ test/module/core.test.js
@@ -1,7 +1,7 @@
 import {describe, it, expect} from 'vitest';
-import {getCommand} from "./src/module/core";
+import {getCommand} from "../../src/js/racing.mjs";
 
-describe('core.mjs', () => {
+describe('racing.mjs', () => {
     describe('getCommand', () => {
         it('4이상의 값이 주어지 면 go를 반환한다.',() => {
             expect(getCommand(4)).to.equal('go');
